Crate mithril_aggregator_client

This crate provides a client to request data from a Mithril Aggregator.

Modules

query

Provides queries to retrieve or send data to a Mithril aggregator

Macros

test_http_compression_is_enabled

Define a test that checks that http compression is enabled for the client.

Structs

AggregatorClientBuilder

A builder of AggregatorHttpClient

AggregatorHttpClient

A client to send HTTP requests to a Mithril Aggregator

Enums

AggregatorHttpClientError

Error structure for the Aggregator Client.

Type Aliases

AggregatorHttpClientResult

Aggregator-client result type