mithril_aggregator_client

Module query

Provides queries to retrieve or send data to a Mithril aggregator

Available queries

Get:

Aggregator: Get current aggregator features, Get current aggregator status

Cardano database v1 (aka Snapshot): List, Get by hash

Cardano database v2: List latest, list for an epoch, list for latest epoch, list for latest epoch with an offset, Get by hash

Cardano stake distribution: List, Get by hash, get for an epoch, get for latest epoch, get for latest epoch with an offset

Cardano transactions snapshot: List, Get by hash

Cardano transactions proof: Get for a list of transactions hashes

Certificate: List, Get by hash, get latest genesis certificate

Epoch settings: Get current epoch settings

Mithril stake distribution: List, Get by hash

Protocol Configuration: Get for a given epoch

Post:

Signature: Register a signature

Registration: Send a signer registration

Cardano database v1 (aka Snapshot): Increment snapshot download statistic

Cardano database v2:

Increment Cardano database snapshot complete or partial restoration statistic

Increment Cardano database snapshot immutables restored statistic

Increment Cardano database snapshot ancillary files restored statistic

Structs

GetAggregatorFeaturesQuery

Query to get the features of the aggregator

GetAggregatorStatusQuery

Query to get the status of the aggregator

GetCardanoDatabaseListQuery

Query to get a list of Cardano database v2 snapshots

GetCardanoDatabaseQuery

Query to get the details of a Cardano database v2 snapshot

GetCardanoStakeDistributionQuery

Query to get a Cardano stake distribution

GetCardanoStakeDistributionsListQuery

Query to get a list of Cardano stake distributions

GetCardanoTransactionProofQuery

Query to get a proof of membership for a list of Cardano transaction hashes

GetCardanoTransactionQuery

Query to get the details of a Cardano transaction snapshot detail

GetCardanoTransactionsListQuery

Query to get a list of Cardano transactions snapshots

GetCertificateQuery

Query to get the details of a certificate

GetCertificatesListQuery

Query to get a list of certificates

GetEpochSettingsQuery

Query to get the current epoch settings

GetMithrilStakeDistributionQuery

Query to get the details of a Mithril stake distribution

GetMithrilStakeDistributionsListQuery

Query to get a list of Mithril stake distributions

GetProtocolConfigurationQuery

Query to get the protocol configuration of a given epoch

GetSnapshotQuery

Query to get the details of a Cardano database v1 snapshot

GetSnapshotsListQuery

Query to get a list of cardano database v1 snapshots

PostIncrementCardanoDatabaseAncillaryRestoredStatisticQuery

Query to notify the aggregator that Cardano database ancillary files have been restored.

PostIncrementCardanoDatabaseImmutablesRestoredStatisticQuery

Query to notify the aggregator that Cardano database immutables files have been restored.

PostIncrementCardanoDatabaseRestorationStatisticQuery

Query to notify the aggregator that a Cardano database restoration has been run.

PostIncrementSnapshotDownloadStatisticQuery

Query to notify the aggregator that a snapshot has been downloaded.

PostRegisterSignatureQuery

Query to register a signature to a Mithril Aggregator.

PostRegisterSignerQuery

Query to register a signer to a Mithril Aggregator.